Implements keys to boost and exclude phrases.
