+++ Source/InputMethodController+CandidateControllerDelegate.swift
@@ -1,0 +1,143 @@
+// Copyright (c) 2022 and onwards The McBopomofo Authors.
+//
+// Permission is hereby granted, free of charge, to any person
+// obtaining a copy of this software and associated documentation
+// files (the "Software"), to deal in the Software without
+// restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following
+// conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+// OTHER DEALINGS IN THE SOFTWARE.
+
+import Cocoa
+import InputMethodKit
+import CandidateUI
+import NotifierUI
+
+extension McBopomofoInputMethodController: CandidateControllerDelegate {
+    func candidateCountForController(_ controller: CandidateController) -> UInt {
+        return if let state = state as? CandidateProvider {
+            UInt(state.candidateCount)
+        } else {
+            0
+        }
+    }
+
+    func candidateController(_ controller: CandidateController, candidateAtIndex index: UInt) -> String {
+        return if let state = state as? CandidateProvider {
+            state.candidate(at: Int(index))
+        } else {
+            ""
+        }
+    }
+
+    func candidateController(_ controller: CandidateController, didSelectCandidateAtIndex index: UInt) {
+        let client = currentClient
+
+        switch state {
+        case let state as InputState.ChoosingCandidate:
+            let selectedCandidate = state.candidates[Int(index)]
+            keyHandler.fixNode(reading: selectedCandidate.reading, value: selectedCandidate.value, originalCursorIndex: Int(state.originalCursorIndex), useMoveCursorAfterSelectionSetting: true)
+
+            guard let inputting = keyHandler.buildInputtingState() as? InputState.Inputting else {
+                return
+            }
+
+            switch keyHandler.inputMode {
+            case .plainBopomofo:
+                keyHandler.clear()
+                let composingBuffer = inputting.composingBuffer
+                handle(state: .Committing(poppedText: composingBuffer), client: client)
+                if Preferences.associatedPhrasesEnabled,
+                   let associatePhrases = keyHandler.buildAssociatedPhrasePlainState(withReading: selectedCandidate.reading, value: selectedCandidate.value, useVerticalMode: state.useVerticalMode) as? InputState.AssociatedPhrasesPlain {
+                    self.handle(state: associatePhrases, client: client)
+                } else {
+                    handle(state: .Empty(), client: client)
+                }
+            case .bopomofo:
+                handle(state: inputting, client: client)
+                if Preferences.associatedPhrasesEnabled {
+                    var textFrame = NSRect.zero
+                    let attributes: [AnyHashable: Any]? = (client as? IMKTextInput)?.attributes(forCharacterIndex: 0, lineHeightRectangle: &textFrame)
+                    let useVerticalMode = (attributes?["IMKTextOrientation"] as? NSNumber)?.intValue == 0 || false
+
+                    let state = keyHandler.buildInputtingState()
+                    keyHandler.handleAssociatedPhrase(with: state, useVerticalMode: useVerticalMode, stateCallback: { newState in
+                        self.handle(state: newState, client: client)
+                    }, errorCallback: {
+                        if (Preferences.beepUponInputError) {
+                            NSSound.beep()
+                        }
+                    }, useShiftKey: true)
+                }
+            default:
+                break
+            }
+        case let state as InputState.AssociatedPhrases:
+            let candidate = state.candidates[Int(index)]
+            keyHandler.fixNodeForAssociatedPhraseWithPrefix(at: state.prefixCursorIndex, prefixReading: state.prefixReading, prefixValue: state.prefixValue, associatedPhraseReading: candidate.reading, associatedPhraseValue: candidate.value)
+            guard let inputting = keyHandler.buildInputtingState() as? InputState.Inputting else {
+                return
+            }
+            handle(state: inputting, client: client)
+            break
+        case let state as InputState.AssociatedPhrasesPlain:
+            let selectedCandidate = state.candidates[Int(index)]
+            handle(state: .Committing(poppedText: selectedCandidate.value), client: currentClient)
+            if Preferences.associatedPhrasesEnabled,
+               let associatePhrases = keyHandler.buildAssociatedPhrasePlainState(withReading: selectedCandidate.reading, value: selectedCandidate.value, useVerticalMode: state.useVerticalMode) as? InputState.AssociatedPhrasesPlain {
+                self.handle(state: associatePhrases, client: client)
+            } else {
+                handle(state: .Empty(), client: client)
+            }
+        case let state as InputState.SelectingFeature:
+            if let nextState = state.nextState(by: Int(index)) {
+                handle(state: nextState, client: client)
+            }
+        case let state as InputState.SelectingDateMacro:
+            let candidate = state.candidate(at: Int(index))
+            if !candidate.isEmpty {
+                let committing = InputState.Committing(poppedText: candidate)
+                handle(state: committing, client: client)
+            }
+        case let state as InputState.SelectingDictionary:
+            let handled = state.lookUp(usingServiceAtIndex: Int(index), state: state) { state in
+                handle(state: state, client: client)
+            }
+            if handled {
+                let previous = state.previousState
+                let candidateIndex = state.selectedIndex
+                handle(state: previous, client: client)
+                if candidateIndex > 0 {
+                    gCurrentCandidateController?.selectedCandidateIndex = UInt(candidateIndex)
+                }
+            }
+        case let state as InputState.ShowingCharInfo:
+            let text = state.menuTitleValueMapping[Int(index)].1
+            NSPasteboard.general.declareTypes([.string], owner: nil)
+            NSPasteboard.general.setString(text, forType: .string)
+            NotifierController.notify(message:String(format:NSLocalizedString("%@ has been copied.", comment: ""), text))
+
+            let previous = state.previousState.previousState
+            let candidateIndex = state.previousState.selectedIndex
+            handle(state: previous, client: client)
+            if candidateIndex > 0 {
+                gCurrentCandidateController?.selectedCandidateIndex = UInt(candidateIndex)
+            }
+        default:
+            break
+        }
+    }
+}
